Correct Blockquote (#14)
Markdown block quotes break on multi-line inputs with an extra space between
them, something someone might type in the area text field on the page before.
We additionally replace new lines in the comments with `<br>` so they show up
correctly in the webpage.

Additionally, we were still using install_requires on MAVirtualCourt instead of
MassAccess. This change shouldn't change anything in the YAML files, checked
that there are no references to MAVirtualCourt.